+++ front/src/texts/texts.json
@@ -7,7 +7,9 @@
     "Supervision": ["Valvonnat", "Supervisions"],
     "Tablet": ["Tablettinäkymä", "Tablet view"],
     "trackCRUD": ["Ratojen päivitys", "Edit tracks"],
-    "SignOut": ["Kirjaudu ulos", "Sign Out"],
+    "SignOut": ["Kirjaudu ulos", "Sign Out"]
+  },
+  "banner": {
     "Notification": ["Sinulla on varmistamattomia päävalvojavuoroja!", "You have unconfirmed range officer reservations!"],
     "Check": ["TARKISTA", "CHECK"]
   },
